What is Facebook advertisements ?
Facebook advertisement is the online advertisements platform. We can create ads and publish on the Facebook platform to be targeted to its users. Facebook advertisements are a type of display ad and can only be seen by the platform's users who have been selected using Facebook's targeting tools. https://www.besttech.in/facebook-advertisement/